requirements.txt
fastapi==0.121.3
uvicorn==0.38.0
pydantic==2.12.4
pydantic[email]==2.12.4
psycopg2-binary==2.9.11
passlib[argon2]==1.7.4
pyjwt==2.8.1
python-dotenv==1.2.1
SQLAlchemy==2.0.44
alembic==1.13.1
